refactor: add context module and service
