+++ schema.graphql
@@ -396,15 +396,14 @@ type UpdateProposalPayload {
 type Query {
   event(id: Int): Event!
   events(input: EventsInput!): [Event!]!
+  post(id: Int!): Post!
   groupMemberRequest(groupId: Int!): GroupMemberRequest
-  groupRole(id: Int!): GroupRole!
-  groupRoles: [GroupRole!]!
   group(id: Int, name: String): Group!
   groups: [Group!]!
   publicGroups: [Group!]!
   publicGroupsFeed: [FeedItem!]!
-  post(id: Int!): Post!
-  proposal(id: Int!): Proposal!
+  groupRole(id: Int!): GroupRole!
+  groupRoles: [GroupRole!]!
   me: User!
   user(id: Int, name: String): User!
   users: [User!]!
@@ -412,6 +411,7 @@ type Query {
   isFirstUser: Boolean!
   serverRole(id: Int!): ServerRole!
   serverRoles: [ServerRole!]!
+  proposal(id: Int!): Proposal!
   authCheck: Boolean!
   serverInvite(token: String!): ServerInvite!
   serverInvites: [ServerInvite!]!
@@ -426,43 +426,43 @@ type Mutation {
   createEvent(eventData: CreateEventInput!): CreateEventPayload!
   updateEvent(eventData: UpdateEventInput!): UpdateEventPayload!
   deleteEvent(id: Int!): Boolean!
-  createGroupMemberRequest(groupId: Int!): CreateGroupMemberRequestPayload!
-  approveGroupMemberRequest(id: Int!): ApproveGroupMemberRequestPayload!
-  cancelGroupMemberRequest(id: Int!): Boolean!
-  denyGroupMemberRequest(id: Int!): Boolean!
-  createGroupRole(groupRoleData: CreateGroupRoleInput!): CreateGroupRolePayload!
-  updateGroupRole(groupRoleData: UpdateGroupRoleInput!): UpdateGroupRolePayload!
-  deleteGroupRole(id: Int!): Boolean!
-  deleteGroupRoleMember(groupRoleMemberData: DeleteGroupRoleMemberInput!): DeleteGroupRoleMemberPayload!
-  createGroup(groupData: CreateGroupInput!): CreateGroupPayload!
-  updateGroup(groupData: UpdateGroupInput!): UpdateGroupPayload!
-  deleteGroup(id: Int!): Boolean!
-  leaveGroup(id: Int!): Boolean!
-  createPost(postData: CreatePostInput!): CreatePostPayload!
-  updatePost(postData: UpdatePostInput!): UpdatePostPayload!
-  deletePost(id: Int!): Boolean!
-  createProposal(proposalData: CreateProposalInput!): CreateProposalPayload!
-  updateProposal(proposalData: UpdateProposalInput!): UpdateProposalPayload!
-  deleteProposal(id: Int!): Boolean!
-  refreshToken: Boolean!
-  updateUser(userData: UpdateUserInput!): UpdateUserPayload!
-  deleteUser(id: Int!): Boolean!
-  followUser(id: Int!): FollowUserPayload!
-  unfollowUser(id: Int!): Boolean!
   createEventAttendee(eventAttendeeData: CreateEventAttendeeInput!): CreateEventAttendeePayload!
   updateEventAttendee(eventAttendeeData: UpdateEventAttendeeInput!): UpdateEventAttendeePayload!
   deleteEventAttendee(eventId: Int!): Boolean!
   deleteImage(id: Int!): Boolean!
+  createPost(postData: CreatePostInput!): CreatePostPayload!
+  updatePost(postData: UpdatePostInput!): UpdatePostPayload!
+  deletePost(id: Int!): Boolean!
   createLike(likeData: CreateLikeInput!): CreateLikePayload!
   deleteLike(likeData: DeleteLikeInput!): Boolean!
+  createGroupMemberRequest(groupId: Int!): CreateGroupMemberRequestPayload!
+  approveGroupMemberRequest(id: Int!): ApproveGroupMemberRequestPayload!
+  cancelGroupMemberRequest(id: Int!): Boolean!
+  denyGroupMemberRequest(id: Int!): Boolean!
+  createGroup(groupData: CreateGroupInput!): CreateGroupPayload!
+  updateGroup(groupData: UpdateGroupInput!): UpdateGroupPayload!
+  deleteGroup(id: Int!): Boolean!
+  leaveGroup(id: Int!): Boolean!
   updateGroupConfig(groupConfigData: UpdateGroupConfigInput!): UpdateGroupPayload!
+  createGroupRole(groupRoleData: CreateGroupRoleInput!): CreateGroupRolePayload!
+  updateGroupRole(groupRoleData: UpdateGroupRoleInput!): UpdateGroupRolePayload!
+  deleteGroupRole(id: Int!): Boolean!
+  deleteGroupRoleMember(groupRoleMemberData: DeleteGroupRoleMemberInput!): DeleteGroupRoleMemberPayload!
+  updateUser(userData: UpdateUserInput!): UpdateUserPayload!
+  deleteUser(id: Int!): Boolean!
+  followUser(id: Int!): FollowUserPayload!
+  unfollowUser(id: Int!): Boolean!
   createServerRole(serverRoleData: CreateServerRoleInput!): CreateServerRolePayload!
   updateServerRole(serverRoleData: UpdateServerRoleInput!): UpdateServerRolePayload!
   deleteServerRole(id: Int!): Boolean!
   deleteServerRoleMember(serverRoleMemberData: DeleteServerRoleMemberInput!): DeleteServerRoleMemberPayload!
+  createProposal(proposalData: CreateProposalInput!): CreateProposalPayload!
+  updateProposal(proposalData: UpdateProposalInput!): UpdateProposalPayload!
+  deleteProposal(id: Int!): Boolean!
   createVote(voteData: CreateVoteInput!): CreateVotePayload!
   updateVote(voteData: UpdateVoteInput!): UpdateVotePayload!
   deleteVote(id: Int!): Boolean!
+  refreshToken: Boolean!
   login(input: LoginInput!): LoginPayload!
   signUp(input: SignUpInput!): SignUpPayload!
   logOut: Boolean!
@@ -497,6 +497,55 @@ input UpdateEventInput {
   endsAt: DateTime
 }
 
+input CreateEventAttendeeInput {
+  eventId: Int!
+  status: String!
+}
+
+input UpdateEventAttendeeInput {
+  eventId: Int!
+  status: String!
+}
+
+input CreatePostInput {
+  groupId: Int
+  eventId: Int
+  body: String
+  images: [Upload!]
+}
+
+input UpdatePostInput {
+  id: Int!
+  body: String
+  images: [Upload!]
+}
+
+input CreateLikeInput {
+  postId: Int
+}
+
+input DeleteLikeInput {
+  postId: Int
+}
+
+input CreateGroupInput {
+  name: String!
+  description: String!
+  coverPhoto: Upload
+}
+
+input UpdateGroupInput {
+  id: Int!
+  name: String
+  description: String
+  coverPhoto: Upload
+}
+
+input UpdateGroupConfigInput {
+  groupId: Int!
+  privacy: String
+}
+
 input CreateGroupRoleInput {
   name: String!
   color: String!
@@ -529,30 +578,40 @@ input DeleteGroupRoleMemberInput {
   userId: Int!
 }
 
-input CreateGroupInput {
+input UpdateUserInput {
+  id: Int!
   name: String!
-  description: String!
+  bio: String!
+  profilePicture: Upload
   coverPhoto: Upload
 }
 
-input UpdateGroupInput {
+input CreateServerRoleInput {
+  name: String!
+  color: String!
+}
+
+input UpdateServerRoleInput {
   id: Int!
   name: String
-  description: String
-  coverPhoto: Upload
+  color: String
+  selectedUserIds: [Int!]
+  permissions: ServerRolePermissionInput
 }
 
-input CreatePostInput {
-  groupId: Int
-  eventId: Int
-  body: String
-  images: [Upload!]
+input ServerRolePermissionInput {
+  manageRoles: Boolean
+  managePosts: Boolean
+  manageComments: Boolean
+  manageEvents: Boolean
+  manageInvites: Boolean
+  createInvites: Boolean
+  removeMembers: Boolean
 }
 
-input UpdatePostInput {
-  id: Int!
-  body: String
-  images: [Upload!]
+input DeleteServerRoleMemberInput {
+  serverRoleId: Int!
+  userId: Int!
 }
 
 input CreateProposalInput {
@@ -588,65 +647,6 @@ input UpdateProposalInput {
   body: String
   action: ProposalActionInput!
   images: [Upload!]
-}
-
-input UpdateUserInput {
-  id: Int!
-  name: String!
-  bio: String!
-  profilePicture: Upload
-  coverPhoto: Upload
-}
-
-input CreateEventAttendeeInput {
-  eventId: Int!
-  status: String!
-}
-
-input UpdateEventAttendeeInput {
-  eventId: Int!
-  status: String!
-}
-
-input CreateLikeInput {
-  postId: Int
-}
-
-input DeleteLikeInput {
-  postId: Int
-}
-
-input UpdateGroupConfigInput {
-  groupId: Int!
-  privacy: String
-}
-
-input CreateServerRoleInput {
-  name: String!
-  color: String!
-}
-
-input UpdateServerRoleInput {
-  id: Int!
-  name: String
-  color: String
-  selectedUserIds: [Int!]
-  permissions: ServerRolePermissionInput
-}
-
-input ServerRolePermissionInput {
-  manageRoles: Boolean
-  managePosts: Boolean
-  manageComments: Boolean
-  manageEvents: Boolean
-  manageInvites: Boolean
-  createInvites: Boolean
-  removeMembers: Boolean
-}
-
-input DeleteServerRoleMemberInput {
-  serverRoleId: Int!
-  userId: Int!
 }
 
 input CreateVoteInput {
